ng-click on master
Since i needed the shopping list id in the clicking function, i had to
move into use ng-click instead of jquery onclick.  the reason for the id
in the fuction is because we might possibly have mutiple lists on a
single page, so multiple masters controlling multiple checkboxes.  I've
now used the i of the shopping list to be inserted into the names of the
checkboxes and sent to the masters ng-click function.