New Views
Added new views.  Planning on adding new dialogs for adding items and
stores, but first will build views and play with a few ideas.  Possible
model change, Items to contain an array of prices and stores.  Another
possibility is Brand, and within brand prices per store.  This will alow
the user to view the possible prices so they can make the best decision
of price to pay.