Store working skeleton
Completed the working skeleton for the stores.  Can now insert a new
store.

The next idea is to get a working skeleton for items working.  The idea
for items model is simply name and descriptions.  But we'll want to have
a different kind of item, possibly named list item or store item, which
will contain the item linked to a store and a price.  That way when
adding an item to a shopping cart, we'll be adding a list itemw hich
will already contain the store and the price.